requirements.txt
Django==4.2.25
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
psycopg2-binary==2.9.10
pytest==8.4.2
pytest-django==4.11.1
gunicorn==23.0.0
PyYAML==6.0.3
uritemplate==4.2.0
django-extensions==3.2.3
django-cors-headers
